requirements.txt
flask==2.2.2
waitress==2.1.2
python-dotenv==0.21.0
sqlmodel==0.0.8
mysqlclient==2.1.0
